pyproject.toml
[tool.poetry]
name = "papershelf"
version = "0.1.0"
description = "A RAG application for storing and querying academic papers"
authors = ["afsharalex <contact@alexafshar.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.1.0"
langgraph = "^0.0.20"
pypdf = "^3.17.0"
sentence-transformers = "^2.2.2"
chromadb = "^0.4.18"
pydantic = "^2.5.2"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
python-dotenv = "^1.0.0"
langchain-openai = "^0.0.2"

[tool.poetry.scripts]
papershelf = "papershelf.main:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
